Make new Problem class to take over responsibility of mathematical problem representation
