+++ src/tespy/networks/network.py
@@ -14,16 +14,13 @@ SPDX-License-Identifier: MIT
 """
 import importlib
 import json
-import math
 import os
 import warnings
 from pathlib import Path
-from time import time
 
 import numpy as np
 import pandas as pd
 from numpy.linalg import norm
-from scipy.optimize import brentq
 from tabulate import tabulate
 
 from tespy.components import CycleCloser
@@ -33,6 +30,7 @@ from tespy.components import WaterElectrolyzer
 from tespy.components.component import component_registry
 from tespy.connections.connection import ConnectionBase
 from tespy.connections.connection import connection_registry
+from tespy.solver import Problem
 from tespy.tools import helpers as hlp
 from tespy.tools import logger
 from tespy.tools.characteristics import CharLine
@@ -196,6 +194,7 @@ class Network:
         self.checked = False
         self.design_path = None
         self.iterinfo = True
+        self._problem = None
         self.units = Units()
 
         msg = 'Default unit specifications:\n'
@@ -708,6 +707,45 @@ class Network:
             )
             raise AttributeError(msg)
 
+    @property
+    def problem(self):
+        """Solver :code:`Problem` instance of the most recent solve call.
+
+        Holds the variable space, equation lookups and the state of the
+        newton algorithm (residual vector, jacobian, increment).
+        """
+        if self._problem is None:
+            msg = (
+                "The network does not hold a problem instance yet. It is "
+                "created with the first call of the solve method."
+            )
+            raise hlp.TESPyNetworkError(msg)
+        return self._problem
+
+    _PROBLEM_ATTRIBUTES = frozenset({
+        "variables_dict", "variable_counter", "residual", "jacobian",
+        "increment", "increment_filter", "lin_dep", "residual_history",
+        "singularity_msg", "iter", "max_iter", "min_iter", "use_cuda",
+        "robust_relax", "oscillation_damping",
+        "num_comp_eq", "num_conn_eq", "num_ude_eq",
+        "get_sorted_residual_index",
+    })
+
+    def __getattr__(self, name):
+        if (
+                name in Network._PROBLEM_ATTRIBUTES
+                and self.__dict__.get("_problem") is not None
+            ):
+            warnings.warn(
+                f"Accessing Network.{name} is deprecated, use "
+                f"Network.problem.{name} instead.",
+                FutureWarning, stacklevel=2
+            )
+            return getattr(self._problem, name)
+        raise AttributeError(
+            f"'{type(self).__name__}' object has no attribute '{name}'"
+        )
+
     def check_topology(self):
         r"""Check if components are connected properly within the network."""
         if len(self.conns) == 0:
@@ -866,12 +904,7 @@ class Network:
         - Set component and connection design point properties.
         - Switch from design/offdesign parameter specification.
         """
-        # keep track of the number of component and connection equations
-        # as well as number of component variables
-        self.num_comp_eq = 0
-        self.num_conn_eq = 0
-        self.variable_counter = 0
-        self.variables_dict = {}
+        self._problem = Problem(self)
 
         # in multiprocessing copies are made of all connections
         # the mass flow branches and fluid branches hold references to
@@ -1142,14 +1175,14 @@ class Network:
                         reference_container.val[f] = fluid0[f]
 
             for fluid in reference_container.is_var:
-                reference_container._J_col[fluid] = self.variable_counter
-                self.variables_dict[self.variable_counter] = {
+                reference_container._J_col[fluid] = self._problem.variable_counter
+                self._problem.variables_dict[self._problem.variable_counter] = {
                     "obj": reference_container,
                     "variable": "fluid",
                     "fluid": fluid,
                     "_represents": linear_dependents["variables"]
                 }
-                self.variable_counter += 1
+                self._problem.variable_counter += 1
 
     def _create_structure_matrix(self):
         self._structure_matrix = {}
@@ -1159,9 +1192,6 @@ class Network:
         self._equation_set_lookup = {}
         self._presolved_equations = []
         self._reference_container_lookup = {}
-        self._equation_lookup = {}
-        self._equation_obj_lookup = {}
-        self._incidence_matrix = {}
 
         num_vars = self._prepare_variables()
 
@@ -1573,15 +1603,15 @@ class Network:
         eq_counter = 0
 
         _eq_counter = self._prepare_network_parts(self.comps["object"], eq_counter)
-        self.num_comp_eq = _eq_counter - eq_counter
+        self._problem.num_comp_eq = _eq_counter - eq_counter
         eq_counter = _eq_counter
 
         _eq_counter = self._prepare_network_parts(self.conns["object"], eq_counter)
-        self.num_conn_eq = _eq_counter - eq_counter
+        self._problem.num_conn_eq = _eq_counter - eq_counter
         eq_counter = _eq_counter
 
         _eq_counter = self._prepare_network_parts(self.user_defined_eq.values(), eq_counter)
-        self.num_ude_eq = _eq_counter - eq_counter
+        self._problem.num_ude_eq = _eq_counter - eq_counter
         eq_counter = _eq_counter
 
     def _prepare_network_parts(self, parts, eq_counter):
@@ -1591,8 +1621,8 @@ class Network:
                 eq_num: (obj.label, eq_name)
                 for eq_num, eq_name in obj._equation_lookup.items()
             }
-            self._equation_lookup.update(eq_map)
-            self._equation_obj_lookup.update(
+            self._problem._equation_lookup.update(eq_map)
+            self._problem._equation_obj_lookup.update(
                 {eq_num: obj for eq_num in obj._equation_lookup}
             )
 
@@ -1600,7 +1630,7 @@ class Network:
                 eq_num: [dependent.J_col for dependent in dependents]
                 for eq_num, dependents in obj._equation_scalar_dependents_lookup.items()
             }
-            self._incidence_matrix.update(dependents_map)
+            self._problem._incidence_matrix.update(dependents_map)
 
             dependents_map = {
                 eq_num: [
@@ -1611,11 +1641,11 @@ class Network:
                 for eq_num, dependents in obj._equation_vector_dependents_lookup.items()
             }
             for eq_num, dependents in dependents_map.items():
-                if eq_num in self._incidence_matrix:
-                    self._incidence_matrix[eq_num] += dependents
+                if eq_num in self._problem._incidence_matrix:
+                    self._problem._incidence_matrix[eq_num] += dependents
 
                 else:
-                    self._incidence_matrix[eq_num] = dependents
+                    self._problem._incidence_matrix[eq_num] = dependents
 
         return eq_counter
 
@@ -1658,14 +1688,14 @@ class Network:
     def _assign_variable_space(self, reference, represents):
         container = reference["object"].get_attr(reference["property"])._reference_container
         if container.is_var:
-            container.J_col = self.variable_counter
-            self.variables_dict[self.variable_counter] = {
+            container.J_col = self._problem.variable_counter
+            self._problem.variables_dict[self._problem.variable_counter] = {
                 "obj": container,
                 "variable": reference["property"],
                 "fluid": None,
                 "_represents": represents
             }
-            self.variable_counter += 1
+            self._problem.variable_counter += 1
 
     def _transform_user_input_to_SI(self):
         """Specification of SI values for user set values."""
@@ -2288,9 +2318,9 @@ class Network:
             list of presolved variables
         """
         represented_variables = []
-        for v in self.variables_dict.values():
+        for v in self._problem.variables_dict.values():
             represented_variables += v["_represents"]
-        if len(self.variables_dict) == 0 and len(self._presolved_equations) == 0:
+        if len(self._problem.variables_dict) == 0 and len(self._presolved_equations) == 0:
             return []
         return [
             (v["object"].label, v["property"])
@@ -2322,7 +2352,7 @@ class Network:
                     self._variable_lookup[v]["property"]
                 ) for v in data["_represents"]
             ]
-            for key, data in self.variables_dict.items()
+            for key, data in self._problem.variables_dict.items()
         }
 
     def print_variables(self):
@@ -2356,7 +2386,7 @@ class Network:
                     self._variable_lookup[v]["property"]
                 ) for v in data["_represents"]
             ]
-            for key, data in self.variables_dict.items()
+            for key, data in self._problem.variables_dict.items()
             if key in number_list
         }
 
@@ -2370,28 +2400,14 @@ class Network:
             parameter defining the equation. In case one parameter defines
             multiple equations, the same equation is repeated.
         """
-        return self._equation_lookup
-
-    def _format_var_label(self, v_idx):
-        v_data = self.variables_dict[v_idx]
-        v_type = v_data["variable"]
-        if v_type == "fluid" and v_data["fluid"] is not None:
-            return f"{v_data['fluid']}{v_idx}"
-        return f"{v_type}{v_idx}"
-
-    @staticmethod
-    def _format_eq_name(eq_name):
-        if isinstance(eq_name, tuple):
-            name, sub_idx = eq_name
-            return f"{name}{{{sub_idx}}}" if sub_idx > 0 else name
-        return eq_name
+        return self._problem._equation_lookup
 
     def print_equations(self):
         """Print a formatted table of equations after presolving."""
         equations = self.get_equations()
         print(f"Equations after presolving ({len(equations)} total):")
         rows = [
-            (eq_num, label, self._format_eq_name(eq_name))
+            (eq_num, label, self._problem._format_eq_name(eq_name))
             for eq_num, (label, eq_name) in sorted(equations.items())
         ]
         if rows:
@@ -2408,24 +2424,24 @@ class Network:
             (variable number, variable type)
         """
         dependencies = {}
-        for eq_idx, dependents in self._incidence_matrix.items():
+        for eq_idx, dependents in self._problem._incidence_matrix.items():
             dependencies.update({
-                self._equation_lookup[eq_idx]:
+                self._problem._equation_lookup[eq_idx]:
                 list(self._get_variables_by_number(dependents).keys())
             })
         return dependencies
 
     def print_equations_with_dependents(self):
         """Print a formatted table of equations and the variables they depend on."""
-        print(f"Equations with dependent variables ({len(self._incidence_matrix)} total):")
+        print(f"Equations with dependent variables ({len(self._problem._incidence_matrix)} total):")
         rows = []
-        for eq_idx, dependents in sorted(self._incidence_matrix.items()):
-            label, eq_name = self._equation_lookup[eq_idx]
+        for eq_idx, dependents in sorted(self._problem._incidence_matrix.items()):
+            label, eq_name = self._problem._equation_lookup[eq_idx]
             dep_str = ", ".join(
-                self._format_var_label(v_idx)
+                self._problem._format_var_label(v_idx)
                 for v_idx, _ in self._get_variables_by_number(dependents).keys()
             )
-            rows.append((eq_idx, label, self._format_eq_name(eq_name), dep_str))
+            rows.append((eq_idx, label, self._problem._format_eq_name(eq_name), dep_str))
         if rows:
             print(tabulate(
                 rows,
@@ -2435,20 +2451,22 @@ class Network:
 
     def print_incidence_matrix(self):
         """Print the incidence matrix with equation rows and variable columns."""
-        eq_indices = sorted(self._incidence_matrix.keys())
+        eq_indices = sorted(self._problem._incidence_matrix.keys())
         all_var_indices = sorted({
             v_idx
-            for deps in self._incidence_matrix.values()
+            for deps in self._problem._incidence_matrix.values()
             for v_idx in deps
         })
 
-        col_labels = [self._format_var_label(v_idx) for v_idx in all_var_indices]
+        col_labels = [
+            self._problem._format_var_label(v_idx) for v_idx in all_var_indices
+        ]
 
         rows = []
         for eq_idx in eq_indices:
-            label, eq_name = self._equation_lookup[eq_idx]
-            row_label = f"{label}.{self._format_eq_name(eq_name)}"
-            dep_set = set(self._incidence_matrix[eq_idx])
+            label, eq_name = self._problem._equation_lookup[eq_idx]
+            row_label = f"{label}.{self._problem._format_eq_name(eq_name)}"
+            dep_set = set(self._problem._incidence_matrix[eq_idx])
             rows.append(
                 [row_label] + ["x" if v in dep_set else "-" for v in all_var_indices]
             )
@@ -2458,13 +2476,16 @@ class Network:
 
     def print_residuals(self):
         """Print a formatted table of equation residuals, sorted by magnitude."""
-        if not hasattr(self, "residual"):
+        if self._problem is None or self._problem.residual is None:
             print("Residuals are not available before the first solve call.")
             return
         rows = []
-        for eq_idx in self.get_sorted_residual_index():
-            label, eq_name = self._equation_lookup[eq_idx]
-            rows.append((eq_idx, label, self._format_eq_name(eq_name), self.residual[eq_idx]))
+        for eq_idx in self._problem.get_sorted_residual_index():
+            label, eq_name = self._problem._equation_lookup[eq_idx]
+            rows.append((
+                eq_idx, label, self._problem._format_eq_name(eq_name),
+                self._problem.residual[eq_idx]
+            ))
         print(f"Residuals per equation ({len(rows)} total, sorted by magnitude):")
         if rows:
             print(tabulate(
@@ -2473,21 +2494,6 @@ class Network:
                 tablefmt="simple",
                 floatfmt=".3e",
             ))
-
-    def _get_equations_by_number(self, number_list) -> dict:
-        """Get the actual equations after presolving the problem by equation
-        number
-
-        Returns
-        -------
-        dict
-            Lookup with equation number as index and tuple of label and
-            parameter defining the equation. In case one parameter defines
-            multiple equations, the same equation is repeated.
-        """
-        return {
-            k: v for k, v in self._equation_lookup.items() if k in number_list
-        }
 
     def get_linear_dependents_by_object(self, obj, prop) -> list:
         """Get the list of linear dependent variables for a specified variable
@@ -2540,16 +2546,6 @@ class Network:
                 variables = [self._variable_lookup[v] for v in dependents["variables"]]
                 return [(v["object"].label, v["property"]) for v in variables]
         raise KeyError(f"Variable index {idx} not found in any dependency group.")
-
-    def get_sorted_residual_index(self) -> list[int]:
-        """Get the sorted array of residual indices.
-
-        Returns
-        -------
-        list[int]
-            List of variable numbers, the index values.
-        """
-        return list(np.argsort(np.abs(self.residual))[::-1])
 
     def solve(self, mode, init_path=None, design_path=None,
               max_iter=50, min_iter=4, init_only=False, init_previous=True,
@@ -2638,13 +2634,7 @@ class Network:
 
         self.init_path = init_path
         self.design_path = design_path
-        self.max_iter = max_iter
-        self.min_iter = min_iter
         self.init_previous = init_previous
-        self.iter = 0
-        self.use_cuda = use_cuda
-        self.robust_relax = robust_relax
-        self.oscillation_damping = oscillation_damping
         self.skip_postprocess = skip_postprocess
 
         if self.skip_postprocess:
@@ -2654,13 +2644,13 @@ class Network:
             )
             logger.debug(msg)
 
-        if self.use_cuda and cu is None:
+        if use_cuda and cu is None:
             msg = (
                 'Specifying use_cuda=True requires cupy to be installed on '
                 'your machine. Numpy will be used instead.'
             )
             logger.warning(msg)
-            self.use_cuda = False
+            use_cuda = False
 
         if mode not in ['offdesign', 'design']:
             msg = 'Mode must be "design" or "offdesign".'
@@ -2689,13 +2679,18 @@ class Network:
 
         self._check_determination()
 
-        n = self.variable_counter
-        self._incidence_matrix_dense = np.zeros((n, n))
-        for row, cols in self._incidence_matrix.items():
-            self._incidence_matrix_dense[row, cols] = 1
+        n = self._problem.variable_counter
+        self._problem._incidence_matrix_dense = np.zeros((n, n))
+        for row, cols in self._problem._incidence_matrix.items():
+            self._problem._incidence_matrix_dense[row, cols] = 1
 
         try:
-            self._solve_loop(print_results=print_results)
+            self._problem.solve_loop(
+                max_iter=max_iter, min_iter=min_iter, use_cuda=use_cuda,
+                robust_relax=robust_relax,
+                oscillation_damping=oscillation_damping,
+                iterinfo=self.iterinfo, print_results=print_results
+            )
         except ValueError as e:
             self.status = 99
             msg = f"Simulation crashed due to an unexpected error:\n{e}"
@@ -2703,17 +2698,18 @@ class Network:
             self._unload_variables()
             return
 
+        self.status = self._problem.status
         self._unload_variables()
 
         if self.status == 3:
-            logger.error(self.singularity_msg)
+            logger.error(self._problem.singularity_msg)
             return
 
         elif self.status == 2:
             msg = (
                 "The solver does not seem to make any progress, aborting "
                 "calculation. Residual value is "
-                "{:.2e}".format(norm(self.residual)) +
+                "{:.2e}".format(norm(self._problem.residual)) +
                 "\nPossible reasons include:\n"
                 " - fluid properties moving outside the valid range of the "
                 "property database (consider adjusting p_range or h_range),\n"
@@ -2731,83 +2727,16 @@ class Network:
         logger.info(msg)
         return
 
-    def _solve_loop(self, print_results=True):
-        r"""Loop of the newton algorithm."""
-        # parameter definitions
-        self.residual_history = np.array([])
-        self.residual = np.zeros([self.variable_counter])
-        self.increment = np.ones([self.variable_counter])
-        self.jacobian = np.zeros((self.variable_counter, self.variable_counter))
-        self._prev_residual = None
-
-        self.start_time = time()
-
-        if self.iterinfo:
-            self._print_iterinfo_head(print_results)
-
-        for self.iter in range(self.max_iter):
-            self.increment_filter = np.absolute(self.increment) < ERR ** 2
-            self._solve_iteration()
-            self.residual_history = np.append(
-                self.residual_history, norm(self.residual)
-            )
-            if self.iterinfo:
-                self._print_iterinfo_body(print_results)
-
-            if self.lin_dep:
-                self.status = 3
-                break
-
-            elif self.iter > 40:
-                if (
-                    all(
-                        self.residual_history[(self.iter - 3):] >= self.residual_history[-3] * 0.95
-                    ) and self.residual_history[-1] >= self.residual_history[-2] * 0.95
-                ):
-                    self.status = 2
-                    break
-
-            if (
-                    self.iter >= self.min_iter - 1
-                    and (self.residual_history[-2:] < ERR ** 0.5).all()
-                    # the increment should also be small, but it does not need
-                    # to be that small
-                    and (abs(self.increment) < ERR ** 0.25).all()
-                ):
-                self.status = 0
-                break
-
-        self.end_time = time()
-
-        if self.iterinfo:
-            self._print_iterinfo_tail(print_results)
-
-        if self.iter == self.max_iter - 1:
-            msg = (
-                f"Reached maximum iteration count ({self.max_iter}), "
-                "calculation stopped. Residual value is "
-                "{:.2e}. ".format(norm(self.residual)) +
-                "\nPossible reasons include:\n"
-                " - fluid properties moving outside the valid range of the "
-                "property database (consider adjusting p_range or h_range),\n"
-                " - an impossible constraint that can never be satisfied \n"
-                " - bad starting values causing the Newton solver to diverge.\n"
-                "Use nw.print_residuals() to identify which equations have "
-                "the largest residuals."
-            )
-            logger.warning(msg)
-            self.status = 2
-
     def _check_determination(self):
         r"""Check, if the number of supplied parameters is sufficient."""
-        msg = f'Number of connection equations: {self.num_conn_eq}.'
+        msg = f'Number of connection equations: {self._problem.num_conn_eq}.'
         logger.debug(msg)
-        msg = f'Number of component equations: {self.num_comp_eq}.'
+        msg = f'Number of component equations: {self._problem.num_comp_eq}.'
         logger.debug(msg)
-        msg = f'Number of user defined equations: {self.num_ude_eq}.'
+        msg = f'Number of user defined equations: {self._problem.num_ude_eq}.'
         logger.debug(msg)
 
-        msg = f'Total number of variables: {self.variable_counter}.'
+        msg = f'Total number of variables: {self._problem.variable_counter}.'
         logger.debug(msg)
 
         _hint = (
@@ -2817,581 +2746,26 @@ class Network:
             "equation depends on, or nw.print_incidence_matrix() for a compact "
             "overview."
         )
-        n = self.num_comp_eq + self.num_conn_eq + self.num_ude_eq
-        if n > self.variable_counter:
+        n = (
+            self._problem.num_comp_eq + self._problem.num_conn_eq
+            + self._problem.num_ude_eq
+        )
+        if n > self._problem.variable_counter:
             msg = (
-                f"You have provided too many parameters: {self.variable_counter} "
+                f"You have provided too many parameters: {self._problem.variable_counter} "
                 f"required, {n} supplied. Aborting calculation!{_hint}"
             )
             logger.error(msg)
             self.status = 12
             raise hlp.TESPyNetworkError(msg)
-        elif n < self.variable_counter:
+        elif n < self._problem.variable_counter:
             msg = (
-                f"You have not provided enough parameters: {self.variable_counter} "
+                f"You have not provided enough parameters: {self._problem.variable_counter} "
                 f"required, {n} supplied. Aborting calculation!{_hint}"
             )
             logger.error(msg)
             self.status = 11
             raise hlp.TESPyNetworkError(msg)
-
-    def _print_iterinfo_head(self, print_results=True):
-        """Print head of convergence progress."""
-        # Start with defining the format here
-        self.iterinfo_fmt = ' {iter:5s} | {residual:10s} | {progress:10s} '
-        self.iterinfo_fmt += '| {massflow:10s} | {pressure:10s} | {enthalpy:10s} '
-        self.iterinfo_fmt += '| {fluid:10s} | {energy:10s} | {component:10s} '
-        # Use the format to create the first logging entry
-        msg = self.iterinfo_fmt.format(
-            iter='iter',
-            residual='residual',
-            progress='progress',
-            massflow='massflow',
-            pressure='pressure',
-            enthalpy='enthalpy',
-            fluid='fluid',
-            energy='energy',
-            component='component'
-        )
-        logger.progress(0, msg)
-        msg2 = '-' * 7 + '+------------' * 8
-
-        logger.progress(0, msg2)
-        if print_results:
-            print('\n' + msg + '\n' + msg2)
-
-    def _print_iterinfo_body(self, print_results=True):
-        """Print convergence progress."""
-        m = [k for k, v in self.variables_dict.items() if v["variable"] == "m"]
-        p = [k for k, v in self.variables_dict.items() if v["variable"] == "p"]
-        h = [k for k, v in self.variables_dict.items() if v["variable"] == "h"]
-        fl = [k for k, v in self.variables_dict.items() if v["variable"] == "fluid"]
-        e = [k for k, v in self.variables_dict.items() if v["variable"] == "E"]
-        cp = [k for k in self.variables_dict if k not in m + p + h + fl + e]
-
-        iter_str = str(self.iter + 1)
-        residual_norm = norm(self.residual)
-        residual = 'NaN'
-        progress = 'NaN'
-        massflow = 'NaN'
-        pressure = 'NaN'
-        enthalpy = 'NaN'
-        fluid = 'NaN'
-        energy = 'NaN'
-        component = 'NaN'
-
-        progress_val = -1
-
-        if not np.isnan(residual_norm):
-            residual = '{:.2e}'.format(residual_norm)
-
-            if not self.lin_dep:
-                massflow = '{:.2e}'.format(norm(self.increment[m]))
-                pressure = '{:.2e}'.format(norm(self.increment[p]))
-                enthalpy = '{:.2e}'.format(norm(self.increment[h]))
-                fluid = '{:.2e}'.format(norm(self.increment[fl]))
-                energy  = '{:.2e}'.format(norm(self.increment[e]))
-                component  = '{:.2e}'.format(norm(self.increment[cp]))
-
-            # This should not be hardcoded here.
-            if residual_norm > np.finfo(float).eps * 100:
-                progress_min = math.log(ERR)
-                progress_max = math.log(ERR ** 0.5) * -1
-                progress_val = math.log(max(residual_norm, ERR)) * -1
-                # Scale to 0-1
-                progress_scaled = (
-                    (progress_val - progress_min)
-                    / (progress_max - progress_min)
-                )
-                progress_val = max(0, min(1, progress_scaled))
-                # Scale to 100%
-                progress_val = int(progress_val * 100)
-            else:
-                progress_val = 100
-
-            progress = '{:d} %'.format(progress_val)
-
-        msg = self.iterinfo_fmt.format(
-            iter=iter_str,
-            residual=residual,
-            progress=progress,
-            massflow=massflow,
-            pressure=pressure,
-            enthalpy=enthalpy,
-            fluid=fluid,
-            energy=energy,
-            component=component
-        )
-        logger.progress(progress_val, msg)
-        if print_results:
-            print(msg)
-
-    def _print_iterinfo_tail(self, print_results=True):
-        """Print tail of convergence progress."""
-        num_iter = self.iter + 1
-        clc_time = self.end_time - self.start_time
-        num_ips = num_iter / clc_time if clc_time > 1e-10 else np.inf
-        msg = '-' * 7 + '+------------' * 7
-        logger.progress(100, msg)
-        msg = (
-            "Total iterations: {0:d}, Calculation time: {1:.2f} s, "
-            "Iterations per second: {2:.2f}"
-        ).format(num_iter, clc_time, num_ips)
-        logger.debug(msg)
-        if print_results:
-            print(msg)
-        return
-
-    def _search_reducing_step(self, row, col):
-        """Find the increment for variable col that reduces equation row's
-        residual.
-
-        Searches both +/- directions with geometrically growing step sizes
-        (x2 per iteration, up to 20 iterations each). Works for both scalar
-        variables (m, h, p, E) and vector variables (fluid mass fractions).
-        Prefers the side that produces a sign change in the residual, which
-        guarantees a root in the bracket [x0, x0±d] by the IVT, and refines
-        its location with Brent's method. If both sides bracket a root, the
-        tighter one (smaller |r| at the probe point) is used. Falls back to a
-        secant step if brentq raises, and to the lower-magnitude heuristic
-        when neither side yields a sign change.
-
-        Returns the step to add to the variable, or None if neither direction
-        improves the residual.
-        """
-        obj = self._equation_obj_lookup.get(row)
-        if obj is None:
-            return None
-        _, (param_name, sub_idx) = self._equation_lookup[row]
-        if param_name not in obj.equations:
-            return None
-        data = obj.equations[param_name]
-
-        var_data = self.variables_dict[col]
-        container = var_data["obj"]
-
-        if var_data["variable"] == "fluid":
-            fluid_key = var_data["fluid"]
-            x0 = container.val[fluid_key]
-            # Maintain sum=1 by adjusting the largest other variable fluid by
-            # the same delta in the opposite direction.
-            other_var_fluids = [f for f in container.is_var if f != fluid_key]
-            if other_var_fluids:
-                companion = max(other_var_fluids, key=lambda f: container.val.get(f, 0))
-                companion_x0 = container.val[companion]
-            else:
-                companion = None
-                companion_x0 = None
-
-            def set_x(v):
-                container.val[fluid_key] = v
-                if companion is not None:
-                    container.val[companion] = companion_x0 - (v - x0)
-        else:
-            x0 = container._val_SI
-
-            def set_x(v):
-                container._val_SI = v
-
-        r0 = self.residual[row]
-        abs_r0 = abs(r0)
-
-        def eval_r(x):
-            set_x(x)
-            try:
-                result = data.func(**data.func_params)
-            except Exception:
-                return None
-            finally:
-                set_x(x0)
-            if hasattr(result, '__iter__'):
-                result = list(result)
-                return result[sub_idx] if sub_idx < len(result) else result[0]
-            return result
-
-        # Guard against x0 == 0 producing a zero initial step
-        d = max(abs(x0) * 0.1, 1e-3)
-        found_plus = None
-        found_minus = None
-
-        for _ in range(20):
-            if found_plus is None:
-                r = eval_r(x0 + d)
-                if r is not None and r != r0:
-                    found_plus = (d, r)
-
-            if found_minus is None:
-                r = eval_r(x0 - d)
-                if r is not None and r != r0:
-                    found_minus = (d, r)
-
-            if found_plus is not None and found_minus is not None:
-                break
-            d *= 2
-
-        plus_sign_change = found_plus is not None and r0 * found_plus[1] < 0
-        minus_sign_change = found_minus is not None and r0 * found_minus[1] < 0
-
-        if plus_sign_change or minus_sign_change:
-            # Both sides bracket a root: prefer the tighter probe (smaller |r|)
-            if plus_sign_change and minus_sign_change:
-                plus_d, plus_r = found_plus
-                minus_d, minus_r = found_minus
-                if abs(plus_r) <= abs(minus_r):
-                    sign, step_d, r_val = +1, plus_d, plus_r
-                else:
-                    sign, step_d, r_val = -1, minus_d, minus_r
-            elif plus_sign_change:
-                sign, step_d, r_val = +1, found_plus[0], found_plus[1]
-            else:
-                sign, step_d, r_val = -1, found_minus[0], found_minus[1]
-
-            a = x0
-            b = x0 + sign * step_d
-            try:
-                tol = max(abs(x0) * 1e-6, 1e-10)
-                x_root = brentq(
-                    eval_r, min(a, b), max(a, b), xtol=tol, maxiter=5
-                )
-                return x_root - x0
-            except Exception:
-                pass
-
-            # Secant fallback: linear interpolation between x0 and the probe
-            return sign * step_d * (-r0) / (r_val - r0)
-
-        # No sign change found - fall back to lower-magnitude direction
-        if found_plus is None and found_minus is None:
-            return None
-        if found_plus is None:
-            step_d, r_val = found_minus
-            return -step_d if abs(r_val) < abs_r0 else None
-        if found_minus is None:
-            step_d, r_val = found_plus
-            return +step_d if abs(r_val) < abs_r0 else None
-
-        plus_d, plus_r = found_plus
-        minus_d, minus_r = found_minus
-        if abs(plus_r) <= abs(minus_r):
-            return +plus_d if abs(plus_r) < abs_r0 else None
-        else:
-            return -minus_d if abs(minus_r) < abs_r0 else None
-
-    def _fill_jacobian_surrogates(self):
-        """Restore invertibility for all-zero rows and find better steps.
-
-        For each row that is entirely zero but expected to have non-zero
-        entries (per the incidence matrix), inserts 1 in the expected positions
-        so the Jacobian can be inverted for all other variables.
-        Subsequently searches value of associated variable(s) to find the
-        increment for the affected variable(s) that reduces that equation's
-        residual.
-
-        Returns a dict {col: step} of increment overrides to apply after the
-        inversion.
-        """
-        overrides = {}
-        for row in self._check_all_zero_rows(self.jacobian):
-            for col in self._incidence_matrix.get(row, []):
-                if self.jacobian[row, col] == 0.0:
-                    self.jacobian[row, col] = 1.0
-                    step = self._search_reducing_step(row, col)
-                    if step is not None:
-                        overrides[col] = step
-        return overrides
-
-    def _invert_jacobian(self):
-        """Compute Newton step, storing result in self.increment. Sets self.lin_dep."""
-        self.lin_dep = False
-        self.increment = self.residual * 0
-        if len(self.variables_dict) == 0:
-            return
-
-        self._check_residual_and_jacobian_for_nan()
-
-        overrides = self._fill_jacobian_surrogates()
-
-        try:
-            if self.use_cuda:
-                self.increment = cu.asnumpy(cu.dot(
-                    cu.linalg.inv(cu.asarray(self.jacobian)),
-                    -cu.asarray(self.residual)
-                ))
-            else:
-                row_scales = np.abs(self.jacobian).max(axis=1)
-                row_scales[row_scales == 0] = 1.0
-                J_eq = self.jacobian / row_scales[:, None]
-
-                col_scales = np.maximum(np.abs(J_eq).max(axis=0), 1e-10)
-                J_sc = J_eq / col_scales[None, :]
-                r_eq = -self.residual / row_scales
-
-                self.increment = np.linalg.solve(J_sc, r_eq) / col_scales
-        except np.linalg.LinAlgError:
-            self.lin_dep = True
-            return
-
-        for col, step in overrides.items():
-            self.increment[col] = step
-
-    def _check_residual_and_jacobian_for_nan(self):
-        """Raise an informative error if a NaN or inf entry is found.
-
-        A single NaN in the residual or Jacobian typically contaminates the
-        whole solution vector once the linear system is solved, so the bad
-        value has to be located here, before the solve, to be able to point
-        at the equation that produced it. Otherwise it propagates silently
-        into the next iteration's variable values and only surfaces several
-        iterations later as an unrelated and confusing error.
-        """
-        nan_rows = set(np.where(~np.isfinite(self.residual))[0])
-        nan_rows |= set(np.where(~np.isfinite(self.jacobian).any(axis=1))[0])
-        if len(nan_rows) == 0:
-            return
-
-        eq_str = ", ".join(
-            f"{lbl}.{self._format_eq_name(eq_name)}"
-            for lbl, eq_name in self._get_equations_by_number(nan_rows).values()
-        )
-        msg = (
-            "The residual or one of the partial derivatives of the "
-            f"following equation(s) is NaN or inf: {eq_str}. Please check "
-            "the inputs to and the implementation of these equations, e.g. "
-            "a UserDefinedEquation."
-        )
-        logger.error(msg)
-        raise hlp.TESPyNetworkError(msg)
-
-    def _diagnose_singularity(self):
-        """Build singularity_msg after a failed matrix solve."""
-        if self.iter == 0 and np.linalg.matrix_rank(self._incidence_matrix_dense) < self._incidence_matrix_dense.shape[0]:
-            self.singularity_msg = (
-                "Detected singularity in Jacobian matrix. This singularity "
-                "is most likely caused by the parametrization of your "
-                "problem and NOT a numerical issue. Double check your "
-                "setup.\n"
-            )
-            self._find_linear_dependencies(self.jacobian)
-            return
-
-        expected_entries = self._incidence_matrix_dense.astype(bool)
-        actual_entries = self.jacobian.astype(bool)
-        rows, cols = np.where(expected_entries != actual_entries)
-
-        missing_entries = []
-        for row, col in zip(rows, cols):
-            lbl, eq_name = self._equation_lookup[row]
-            eq_str = f"{lbl}.{self._format_eq_name(eq_name)}"
-            var_str = self._format_var_label(col)
-            missing_entries += [f"{eq_str}: {var_str}"]
-
-        entries_str = ", ".join(missing_entries)
-        self.singularity_msg = (
-            "Found singularity in Jacobian matrix, calculation aborted! "
-            "The setup of your problem seems to be solvable. It failed "
-            "due to partial derivatives in the Jacobian being zero where "
-            "a non-zero was expected, or vice versa. This usually lies in "
-            "starting value selection or bad convergence.\n"
-            "  The following equation/variable pairs may have an "
-            f"unexpected zero/non-zero partial derivative:  {entries_str}\n"
-        )
-        self._find_linear_dependencies(self.jacobian)
-
-    def _find_linear_dependencies(self, matrix):
-        all_zero_cols = self._check_all_zero_columns(matrix)
-        all_zero_rows = self._check_all_zero_rows(matrix)
-        if len(all_zero_cols) + len(all_zero_rows) == 0:
-            eq_indices = self._cauchy_schwarz_inequality(matrix)
-            eq_str = ", ".join(
-                f"{lbl}.{self._format_eq_name(eq_name)}"
-                for lbl, eq_name in self._get_equations_by_number(eq_indices).values()
-            )
-            self.singularity_msg += (
-                "The following equations form a linear dependency:\n"
-                f"  {eq_str}\n"
-            )
-        else:
-            if len(all_zero_cols) > 0:
-                var_str = ", ".join(self._format_var_label(i) for i in all_zero_cols)
-                self.singularity_msg += (
-                    "The following variables are not associated with any equation:\n"
-                    f"  {var_str}\n"
-                )
-            if len(all_zero_rows) > 0:
-                eq_str = ", ".join(
-                    f"{lbl}.{self._format_eq_name(eq_name)}"
-                    for lbl, eq_name in self._get_equations_by_number(all_zero_rows).values()
-                )
-                self.singularity_msg += (
-                    "The following equations do not depend on any variable:\n"
-                    f"  {eq_str}\n"
-                )
-
-    def _check_all_zero_columns(self, matrix):
-        return np.where((matrix == 0).all(axis=0))[0]
-
-    def _check_all_zero_rows(self, matrix):
-        return np.where((matrix == 0).all(axis=1))[0]
-
-    def _cauchy_schwarz_inequality(self, matrix):
-        n = matrix.shape[0]
-        dependent_equations = []
-        for i in range(n):
-            for j in range(n):
-                if i != j:
-                    inner_product = np.inner(
-                        matrix[i,:],
-                        matrix[j,:]
-                    )
-                    norm_i = np.linalg.norm(matrix[i,:])
-                    norm_j = np.linalg.norm(matrix[j,:])
-
-                    if np.abs(inner_product - norm_j * norm_i) < 1e-5:
-                        dependent_equations += [i]
-        return list(set(dependent_equations))
-
-    def _dampen_oscillating_increments(self):
-        """Halve increments for variables whose residuals changed sign since the last iteration.
-
-        A sign change means the Newton step overshot the zero of that equation.
-        Halving the dependent columns' increments keeps the bracket from growing
-        and makes subsequent iterations behave like bisection for those variables.
-        """
-        if self._prev_residual is None:
-            return
-        sign_flips = np.where(self._prev_residual * self.residual < 0)[0]
-        for row in sign_flips:
-            cols = np.nonzero(self.jacobian[row, :])[0]
-            self.increment[cols] *= 0.5
-
-    def _update_variables(self):
-        # cast dtype to float from numpy float64
-        # this is necessary to keep the doctests running and note make them
-        # look ugly all over the place
-        # I have yet to come up with a better idea, or vectorize all operations
-        # which requires major changes in tespy
-        increment = [float(val) for val in self.increment]
-        # the J_cols here point to actual variables, no need to call to
-        # get_J_col yet
-        relax = 1
-        if self.robust_relax:
-            relax = 0.05 + 0.95 * min(1, self.iter / (0.25 * self.max_iter))
-
-        for _, data in self.variables_dict.items():
-            if data["variable"] in ["m", "h", "E"]:
-                container = data["obj"]
-                container._val_SI += increment[container.J_col] * relax
-            elif data["variable"] in ["p"]:
-                container = data["obj"]
-                p_relax = max(
-                    1, -2 * increment[container.J_col] / container.val_SI
-                )
-                container._val_SI += increment[container.J_col] / p_relax
-            elif data["variable"] == "fluid":
-                container = data["obj"]
-                inc = increment[container.J_col[data["fluid"]]]
-                value = container.val[data["fluid"]]
-                if value > ERR:
-                    f_relax = max(1, -2 * inc / value)
-                else:
-                    f_relax = 1
-
-                container.val[data["fluid"]] += inc / f_relax
-
-                if container.val[data["fluid"]] < ERR:
-                    container.val[data["fluid"]] = 0
-                elif container.val[data["fluid"]] > 1 - ERR:
-                    container.val[data["fluid"]] = 1
-            else:
-                # add increment
-                data["obj"]._val_SI += increment[data["obj"].J_col] * relax
-
-                # keep value within specified value range
-                if data["obj"].val_SI < data["obj"].min_val:
-                    data["obj"].val_SI = data["obj"].min_val
-                elif data["obj"].val_SI > data["obj"].max_val:
-                    data["obj"].val_SI = data["obj"].max_val
-
-    def _adapt_to_variable_bounds(self):
-
-        # this could be in a different place, its kind of in between
-        # network and connection
-        if self.iter < 10:
-            for data in self.variables_dict.values():
-                if type(data["obj"]) == dc_vecvar:
-                    total_mass_fractions = sum(data["obj"].val.values())
-                    for fluid in data["obj"].is_var:
-                        data["obj"]._val[fluid] /= total_mass_fractions
-
-        if norm(self.increment) > 1e-1:
-            for c in self.conns['object']:
-                # check the fluid properties for physical ranges
-                c._adjust_to_property_limits(self)
-
-            for cp in self.comps['object']:
-                cp._adjust_to_property_limits()
-
-        # second check based on component heuristics
-        # - for first three iterations
-        # - only if the increment is sufficiently large
-        # - only in design case
-        if (
-                self.iter < 3
-                and norm(self.increment) > 1e-1
-                and self.mode == "design"
-            ):
-            for cp in self.comps['object']:
-                cp.convergence_check()
-
-            for c in self.conns['object']:
-                c._adjust_to_property_limits(self)
-
-    def _solve_iteration(self):
-        r"""
-        Control iteration step of the newton algorithm.
-
-        - Calculate the residual value for each equation
-        - Calculate the jacobian matrix
-        - Calculate new values for variables
-        - Restrict fluid properties to value ranges
-        - Check component parameters for consistency
-        """
-        self._solve_equations()
-        self._invert_jacobian()
-
-        if self.lin_dep:
-            self._diagnose_singularity()
-            return
-
-        if self.oscillation_damping:
-            self._dampen_oscillating_increments()
-
-        self._update_variables()
-        self._adapt_to_variable_bounds()
-        self._prev_residual = self.residual.copy()
-
-    def _solve_equations(self):
-        r"""
-        Calculate the residual and derivatives of all equations.
-        """
-        to_solve = (
-            self.comps["object"].tolist()
-            + self.conns["object"].tolist()
-            + list(self.user_defined_eq.values())
-        )
-        for obj in to_solve:
-            hlp.solve(obj, self.increment_filter)
-            if len(obj.jacobian) > 0:
-                rows = list(obj.residual.keys())
-                data = list(obj.residual.values())
-                self.residual[rows] = data
-
-                rows = [k[0] for k in obj.jacobian]
-                columns = [k[1] for k in obj.jacobian]
-                data = list(obj.jacobian.values())
-                self.jacobian[rows, columns] = data
-
-            obj.it += 1
 
     def _postprocess(self):
         r"""Calculate connection and component parameters."""
@@ -3852,7 +3226,7 @@ class Network:
         >>> imported_nwk = Network.from_json('exported_nwk.json')
         >>> imported_nwk.iterinfo = False
         >>> imported_nwk.solve('design')
-        >>> imported_nwk.lin_dep
+        >>> imported_nwk.problem.lin_dep
         False
         >>> round(imported_nwk.get_conn('c01').m.val_SI, 1) == mass_flow
         True
