+++ src/tespy/solver/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8
+
+from .problem import Problem  # noqa: F401
